adding manage Remainder

+++ Server/DAL/Services/AvailableAppointmentsDal.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using DAL.Api;
 using DAL.Models;
+using Exeptions;
 using Microsoft.EntityFrameworkCore;
 namespace DAL.Services
 {
@@ -91,9 +92,20 @@ namespace DAL.Services
 
         public async Task<List<AvailableAppointment>> RemoveAllAppointmentsByDateAndTherapist(int therapistId, DateOnly date)
         {
+            var therapistExists = await _DB_Manager.Therapists.FirstOrDefaultAsync(c => c.TherapistId == therapistId);
+            if (therapistExists == null)
+                throw new DALNotFoundException("Therapist details were wrong.");
+
+            DateTime dateStart = date.ToDateTime(TimeOnly.MinValue); // 2025-07-29 00:00:00
+            DateTime dateEnd = date.ToDateTime(TimeOnly.MaxValue);   // 2025-07-29 23:59:59.9999999
+
             var appointments = await _DB_Manager.AvailableAppointments
-                .Where(a => a.TherapistId == therapistId && a.AppointmentDate.Date == date.ToDateTime(TimeOnly.MinValue).Date)
+                .Where(c =>
+                    c.AppointmentDate >= dateStart &&
+                    c.AppointmentDate <= dateEnd &&
+                        c.TherapistId == therapistExists.Id)
                 .ToListAsync();
+
 
             _DB_Manager.AvailableAppointments.RemoveRange(appointments);
             await _DB_Manager.SaveChangesAsync();
